Make sklearn checks work with NumPy < 1.22 (convert object arrays to float for np_nanquantile)

+++ mapie/regression.py
@@ -691,7 +691,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         y_pred_low = np.column_stack(
             [
                 np_nanquantile(
-                    lower_bounds,
+                    lower_bounds.astype(float),
                     _alpha,
                     axis=1,
                     method="lower",
@@ -702,7 +702,7 @@ class MapieRegressor(BaseEstimator, RegressorMixin):
         y_pred_up = np.column_stack(
             [
                 np_nanquantile(
-                    upper_bounds,
+                    upper_bounds.astype(float),
                     1 - _alpha,
                     axis=1,
                     method="higher",
